Split storable.go & typeinfo.go into smaller files
This change makes the code easier to read and maintain.

This commit splits typeinfo.go into:
- compactmap_extradata.go
- extradata.go
- typeinfo.go

This commit also moves:
- SlabIDStorable to slab_id_storable.go
- CBOR tag numbers allocated for atree to cbor_tag_nums.go
- etc
