+++ array.go
@@ -385,70 +385,6 @@ func (a *Array) Set(index uint64, value Value) (Storable, error) {
 	return existingStorable, nil
 }
 
-// uninlineStorableIfNeeded uninlines given storable if needed, and
-// returns uninlined Storable and its ValueID.
-// If given storable is a WrapperStorable, this function uninlines
-// wrapped storable if needed and returns a new WrapperStorable
-// with wrapped uninlined storable and its ValidID.
-func uninlineStorableIfNeeded(storage SlabStorage, storable Storable) (Storable, ValueID, bool, error) {
-	if storable == nil {
-		return storable, emptyValueID, false, nil
-	}
-
-	switch s := storable.(type) {
-	case ArraySlab: // inlined array slab
-		err := s.Uninline(storage)
-		if err != nil {
-			return nil, emptyValueID, false, err
-		}
-
-		slabID := s.SlabID()
-
-		newStorable := SlabIDStorable(slabID)
-		valueID := slabIDToValueID(slabID)
-
-		return newStorable, valueID, true, nil
-
-	case MapSlab: // inlined map slab
-		err := s.Uninline(storage)
-		if err != nil {
-			return nil, emptyValueID, false, err
-		}
-
-		slabID := s.SlabID()
-
-		newStorable := SlabIDStorable(slabID)
-		valueID := slabIDToValueID(slabID)
-
-		return newStorable, valueID, true, nil
-
-	case SlabIDStorable: // uninlined slab
-		valueID := slabIDToValueID(SlabID(s))
-
-		return storable, valueID, false, nil
-
-	case WrapperStorable:
-		unwrappedStorable := unwrapStorable(s)
-
-		// Uninline wrapped storable if needed.
-		uninlinedWrappedStorable, valueID, uninlined, err := uninlineStorableIfNeeded(storage, unwrappedStorable)
-		if err != nil {
-			return nil, emptyValueID, false, err
-		}
-
-		if !uninlined {
-			return storable, valueID, uninlined, nil
-		}
-
-		// Create a new WrapperStorable with uninlinedWrappedStorable
-		newStorable := s.WrapAtreeStorable(uninlinedWrappedStorable)
-
-		return newStorable, valueID, uninlined, nil
-	}
-
-	return storable, emptyValueID, false, nil
-}
-
 func (a *Array) set(index uint64, value Value) (Storable, error) {
 	existingStorable, err := a.root.Set(a.Storage, a.Address(), index, value)
 	if err != nil {
